TitleScreen.h and TitleScreen.cpp files added

+++ JobHuntingProduction/Player.h
@@ -154,7 +154,7 @@ protected:
 	float rotateSpeed = 450.0f; // Currently using delta time; 7.5f using frame speed
 	float speed = 60.0f; // Currently using delta time; 1.0f using frame speed
 	float sprintSpeed = 120.0f; // 2.0f using frame speed
-	int attackTime = 0;
+	float attackTime = 0.0f;
 
 public:
 	float stamina = 100.0f;
